+++ src/app/api/identity/route.ts
@@ -1,11 +1,12 @@
 import { NextRequest, NextResponse } from 'next/server';
 import { getToken } from 'next-auth/jwt';
+import {IdentityService} from '@/service/identity-service';
 
 export async function GET(req: NextRequest) {
   try {
 
     const token = await getToken({ req, secret: process.env.NEXTAUTH_SECRET });
-    if(!token) {
+    if (!token) {
       return NextResponse.json('No token found');
     }
 
@@ -18,7 +19,12 @@ export async function GET(req: NextRequest) {
       },
     });
     console.log('data:', token);
-  // console.log('data:', data);
+    // console.log('data:', data);
+
+    if (token.error === 'RefreshAccessTokenError') {
+      IdentityService.signOut(req);
+      return NextResponse.json({ error: 'RefreshAccessTokenError' });
+    }
 
     return NextResponse.json(data);
   } catch (error) {