+++ src/components/identity/auth-menu.tsx
@@ -1,12 +1,20 @@
 "use client";
 
 import { useEffect, useRef, useState } from 'react';
-import { Dialog, DialogContent, DialogContentText, DialogTitle, IconButton } from '@mui/material';
+import { IconButton } from '@mui/material';
 import ManageAccountsRoundedIcon from '@mui/icons-material/ManageAccountsRounded';
 import { signIn, signOut, useSession } from 'next-auth/react';
 import BnButton from '../common/bn-button';
+import { basicFetch } from '@/app/lib/fetchFunctions';
+import SignInDialog from './signin-dialog';
+import { fetchService } from '@/service/fetch-service';
 
-const AuthMenu = ({ dict }: { dict: Record<string, string> }) => {
+const AuthenticationMenu = ({ dict }: { dict: Record<string, string> }) => {
+
+    const [isClient, setIsClient] = useState(false);
+    useEffect(() => {
+        setIsClient(true);
+    }, []);
 
     const [menuOpen, setMenuOpen] = useState(false);
     const menuRef = useRef(null);
@@ -35,82 +43,49 @@ const AuthMenu = ({ dict }: { dict: Record<string, string> }) => {
 
 
     const [dialogOpen, setDialogOpen] = useState(false);
-    const { data: session, status } = useSession()
-    const [loading, setLoading] = useState(false);
-    const [loginError, setLoginError] = useState('')
+    const { data: session, status } = useSession()  
 
     useEffect(() => {
-        if (session) {
+        console.log('session:', session);
+        if (session && session.user) {
             console.log('session.user:', session.user?.name);
         }
     }, [session]);
 
-    const handleSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
-        event.preventDefault();
-
-        const formData = new FormData(event.currentTarget);
-        const username = formData.get('username') as string;
-        const password = formData.get('password') as string;
-        const result = await signIn('bnprovider', {
-            redirect: false,
-            username,
-            password,
-        });
-
-        console.log('Sign in result:', result);
-
-        if (result?.error) {
-            console.error('Error during sign in:', result.error);
-            setLoginError('Login failed. Invalid credentials');
-        } else {
-            console.log('Sign in successful:', result);
-            setDialogOpen(false);
-        }
+    const closeDialog = () => {
+        setDialogOpen(false);
     };
 
     const handleSignOut = async () => {
         try {
-            // Define the endpoint URL
-            const endpoint = '/api/identity/signout';
 
-            // Make the POST request
-            const response = await fetch(endpoint, {
-                method: 'POST',
-                headers: {
-                    'Content-Type': 'application/json',
-                },
-                body: JSON.stringify({}),
-            });
-
-            console.log('Sign out response:', response);
-
-            // Check if the response is successful
-            if (!response.ok) {
-                throw new Error('Failed to sign out');
-            }
-
-            // Handle the response (optional)
-            const data = await response.json();
-            console.log('Sign out response:', data);
-
-            // Call the signOut function
+          const result =   await fetchService.nextJsPost('/api/identity/signout', {});     
+          console.log('Sign out result:', result);    
             await signOut();
         } catch (error) {
             console.error('Error during sign out:', error);
         }
     };
 
-
+    useEffect(() => {
+        const testAuth = async () => {
+            const res = await basicFetch<any>(`/api/identity`);
+            if (res.error && res.error === 'RefreshAccessTokenError') {
+                await signOut();
+            }
+        }
+        testAuth();
+    }, []);
 
     return (
         <>
-            <div className="relative inline-block">
+            <div className="relative inline-block mr-2">
                 <IconButton aria-label="language" color="primary" onClick={toggleMenu}>
                     <ManageAccountsRoundedIcon className='text-slate-50' />
-                    <span className='text-sm text-slate-50 ml-1'>{session && session.user?.name}</span>
+                    <span className='text-sm text-slate-50 ml-1 min-w-fit'>{session && session.user?.name}</span>
                 </IconButton>
 
-                {menuOpen && (
+                {isClient && menuOpen && (
                     <div ref={menuRef} className="absolute left-[-20px] mt-0 w-36 bg-slate-800 border border-gray-200 rounded shadow-lg z-50">
                         {status === 'authenticated' ? (
                             <ul className="flex flex-col p-2">
@@ -143,70 +118,9 @@ const AuthMenu = ({ dict }: { dict: Record<string, string> }) => {
                 )}
             </div>
 
-            <Dialog
-                open={dialogOpen}
-                onClose={() => setDialogOpen(false)} >
-                <DialogTitle>{dict.AUTH_welcome}</DialogTitle>
-                <DialogContent>
-                    <DialogContentText>
-                        {dict.AUTH_login}
-                    </DialogContentText>
-                    <form onSubmit={handleSubmit} className='my-2'>
-                        <div className='flex flex-col'>
-                            <label>Username</label>
-                            <input type="text" className='border border-slate-400 w-full p-1' name="username" required onFocus={() => setLoginError('')} />
-                            <label>{dict.AUTH_password}</label>
-                            <input className='border border-slate-400 w-full p-1'
-                                type="password"
-                                name="password"
-                                required />
-
-                            <BnButton type='submit' label={dict.AUTH_login} />
-                        </div>
-                    </form>
-
-                    <div className='text-red-800'>
-                        {loginError}
-                    </div>
-                    <div>
-                        {dict.AUTH_noaccount} <a href="/auth/account" className='text-blue-500'>{dict.AUTH_register}</a>
-                    </div>
-                </DialogContent>
-
-            </Dialog>
-            <Dialog
-                open={dialogOpen}
-                onClose={() => setDialogOpen(false)} >
-                <DialogTitle>{dict.AUTH_welcome}</DialogTitle>
-                <DialogContent>
-                    <DialogContentText>
-                        {dict.AUTH_login}
-                    </DialogContentText>
-                    <form onSubmit={handleSubmit} className='my-2'>
-                        <div className='flex flex-col'>
-                            <label>Username</label>
-                            <input type="text" className='border border-slate-400 w-full p-1' name="username" required onFocus={() => setLoginError('')} />
-                            <label>{dict.AUTH_password}</label>
-                            <input className='border border-slate-400 w-full p-1'
-                                type="password"
-                                name="password"
-                                required />
-
-                            <BnButton type='submit' label={dict.AUTH_login} />
-                        </div>
-                    </form>
-
-                    <div className='text-red-800'>
-                        {loginError}
-                    </div>
-                    <div>
-                        {dict.AUTH_noaccount} <a href="/auth/account" className='text-blue-500'>{dict.AUTH_register}</a>
-                    </div>
-                </DialogContent>
-
-            </Dialog>
+            <SignInDialog dict={dict} isOpen={dialogOpen} closeDialog={closeDialog} /> 
         </>
     );
 }
 
-export default AuthMenu;
+export default AuthenticationMenu;