+++ src/components/session-provider-wrapper.tsx
@@ -3,14 +3,14 @@
 'use client';
 
 import { SessionProvider } from 'next-auth/react';
-import { ReactNode } from 'react';
+import { ReactNode, useEffect, useState } from 'react';
 
 interface Props {
   children: ReactNode;
 }
 
 const SessionProviderWrapper = ({ children }: Props) => {
-  return <SessionProvider>{children}</SessionProvider>;
+  return <SessionProvider>{children}</SessionProvider>
 };
 
 export default SessionProviderWrapper;